Refactor. Gateway Authorization flow
Changes added according to PR comments.
