+++ src/utilities/GenericResponse.js
@@ -1,4 +1,4 @@
-import registry from '../registry/registry.json';
+// import registry from '../registry/registry.json';
 import LoggingService from '../services/LoggingService';
 
 const logger = LoggingService.getLogger('GenericResponse');
@@ -19,18 +19,17 @@ const sendErrorWithAuthorization = (res) => {
   res.status(401).send('Error');
 };
 
-const getPublicKey = async (subscriberId) => {
-  const subscriber = registry.filter(
-    (entry) => entry.subscriber_id === subscriberId,
-  );
-  logger.debug(`subscriber ${JSON.stringify(subscriber)}`);
-  const publicKey = `${subscriber[0].signing_public_key}`;
-  logger.debug(`public Key: ${publicKey}`);
-  return publicKey;
-};
+// const getPublicKey = async (subscriberId) => {
+//   const subscriber = registry.filter(
+//     (entry) => entry.subscriber_id === subscriberId,
+//   );
+//   logger.debug(`subscriber ${JSON.stringify(subscriber)}`);
+//   const publicKey = `${subscriber[0].signing_public_key}`;
+//   logger.debug(`public Key: ${publicKey}`);
+//   return publicKey;
+// };
 
 export default {
   sendErrorWithAuthorization,
   sendAcknowledgement,
-  getPublicKey,
 };
